chore(commands): share CommonArgs struct
